refactor(react): migrate next-13-app-router playground to v8 API (App Router, Next 13)
- Replace storyblokInit/apiPlugin with createApiClient + createRegistry
- Remove StoryblokProvider — no longer needed
- Replace StoryblokStory/StoryblokServerComponent with StoryblokComponent
- Rename blok->block, SbBlokData->BlockContent across all components
- Remove @storyblok/react/rsc subpath imports
