+++ packages/react/playground/next-13-app-router/app/layout.tsx
@@ -1,19 +1,12 @@
-import StoryblokProvider from "@/components/StoryblokProvider";
-
 export const metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "Storyblok Next.js 13 App Router",
+  description: "Storyblok Next.js 13 App Router playground",
 };
-interface RootLayoutType {
-  children: React.ReactNode;
-}
 
-export default function RootLayout({ children }: RootLayoutType) {
+export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
-    <StoryblokProvider>
-      <html lang="en">
-        <body>{children}</body>
-      </html>
-    </StoryblokProvider>
+    <html lang="en">
+      <body>{children}</body>
+    </html>
   );
 }
