+++ packages/react/playground/next-13-app-router/components/Page.tsx
@@ -1,9 +1,11 @@
-import { storyblokEditable, StoryblokServerComponent } from "@storyblok/react/rsc";
+import type { BlockContent } from "@storyblok/react";
+import { storyblokEditable } from "@storyblok/react";
+import { StoryblokComponent } from "@/lib/storyblok";
 
-const Page = ({ blok }) => (
-  <main {...storyblokEditable(blok)}>
-    {blok.body.map((nestedBlok) => (
-      <StoryblokServerComponent blok={nestedBlok} key={nestedBlok._uid} />
+const Page = ({ block }: { block: BlockContent & { body?: BlockContent[] } }) => (
+  <main {...storyblokEditable(block)}>
+    {(block.body ?? []).map((nested) => (
+      <StoryblokComponent block={nested} key={nested._uid} />
     ))}
   </main>
 );
