+++ packages/react/playground/next-13-app-router/components/Teaser.tsx
@@ -1,11 +1,10 @@
-import { storyblokEditable } from "@storyblok/react/rsc";
+import type { BlockContent } from "@storyblok/react";
+import { storyblokEditable } from "@storyblok/react";
 
-const Teaser = ({ blok }) => {
-  return (
-    <h2 data-cy="teaser" {...storyblokEditable(blok)}>
-      {blok.headline}
-    </h2>
-  );
-};
+const Teaser = ({ block }: { block: BlockContent & { headline?: string } }) => (
+  <h2 data-cy="teaser" {...storyblokEditable(block)}>
+    {block.headline}
+  </h2>
+);
 
 export default Teaser;
