+++ packages/react/playground/next-13-app-router/lib/storyblok.ts
@@ -1,10 +1,13 @@
+import { createApiClient, createRegistry } from "@storyblok/react";
 import Page from "@/components/Page";
 import Teaser from "@/components/Teaser";
-import { apiPlugin, storyblokInit } from "@storyblok/react/rsc";
 
-export const getStoryblokApi = storyblokInit({
+// https://app.storyblok.com/#/me/spaces/147897
+export const client = createApiClient({
   accessToken: "OurklwV5XsDJTIE1NJaD2wtt",
-  use: [apiPlugin],
+});
+
+export const { StoryblokComponent } = createRegistry({
   components: {
     teaser: Teaser,
     page: Page,
